+++ cmd/vegabenchmark/setup.go
@@ -171,6 +171,7 @@ func setupVega() (*processor.App, processor.Stats, error) {
 		collateral,
 		oraclesM,
 		broker,
+		stateVarEngine,
 	)
 
 	netParams := netparams.New(log, netparams.NewDefaultConfig(), broker)