+++ execution/market.go
@@ -29,6 +29,7 @@ import (
 	"code.vegaprotocol.io/vega/settlement"
 	"code.vegaprotocol.io/vega/types"
 	"code.vegaprotocol.io/vega/types/num"
+	"code.vegaprotocol.io/vega/types/statevar"
 
 	"github.com/golang/protobuf/proto"
 )
@@ -232,7 +233,8 @@ type Market struct {
 	lastEquityShareDistributed time.Time
 	equityShares               *EquityShares
 
-	stateChanged bool
+	stateVarEngine StateVarEngine
+	stateChanged   bool
 }
 
 // SetMarketID assigns a deterministic pseudo-random ID to a Market.
@@ -276,6 +278,7 @@ func NewMarket(
 	broker Broker,
 	idgen *IDgenerator,
 	as *monitor.AuctionState,
+	stateVarEngine StateVarEngine,
 ) (*Market, error) {
 	if len(mkt.ID) == 0 {
 		return nil, ErrEmptyMarketID
@@ -307,6 +310,7 @@ func NewMarket(
 		now.UnixNano(),
 		mkt.ID,
 		asset,
+		stateVarEngine,
 	)
 
 	settleEngine := settlement.New(
@@ -385,6 +389,7 @@ func NewMarket(
 		lastMidBuyPrice:    num.Zero(),
 		lastBestBidPrice:   num.Zero(),
 		stateChanged:       true,
+		stateVarEngine:     stateVarEngine,
 	}
 
 	market.tradableInstrument.Instrument.Product.NotifyOnTradingTerminated(market.tradingTerminated)
@@ -398,6 +403,12 @@ func appendBytes(bz ...[]byte) []byte {
 		out = append(out, b...)
 	}
 	return out
+}
+
+// UpdateRiskFactorsForTest is a hack for setting the risk factors for tests directly rather than through the consensus engine.
+// Never use this for anything functional.
+func (m *Market) UpdateRiskFactorsForTest() {
+	m.risk.CalculateRiskFactorsForTest()
 }
 
 func (m *Market) Hash() []byte {
@@ -619,9 +630,7 @@ func (m *Market) OnChainTimeUpdate(ctx context.Context, t time.Time) bool {
 	// check auction, if any
 	m.checkAuction(ctx, t)
 
-	// TODO(): handle market start time
-
-	m.risk.CalculateFactors(ctx, t)
+	// TODO(): handle market start time (is this comment still relevant?)
 	timer.EngineTimeCounterAdd()
 
 	m.updateMarketValueProxy()
@@ -803,6 +812,20 @@ func (m *Market) EnterAuction(ctx context.Context) {
 		m.broker.Send(events.NewMarketUpdatedEvent(ctx, *m.mkt))
 		m.stateChanged = true
 	}
+}
+
+// OnOpeningAuctionFirstUncrossingPrice is triggered when the opening auction sees an uncrossing price for the first time and emits
+// an event to the state variable engine.
+func (m *Market) OnOpeningAuctionFirstUncrossingPrice() {
+	asset, _ := m.mkt.GetAsset()
+	m.stateVarEngine.ReadyForTimeTrigger(asset, m.mkt.ID)
+	m.stateVarEngine.NewEvent(asset, m.mkt.ID, statevar.StateVarEventTypeOpeningAuctionFirstUncrossingPrice)
+}
+
+// OnAuctionEnded is called whenever an auction is ended and emits an event to the state var engine.
+func (m *Market) OnAuctionEnded() {
+	asset, _ := m.mkt.GetAsset()
+	m.stateVarEngine.NewEvent(asset, m.mkt.ID, statevar.StateVarEventTypeAuctionEnded)
 }
 
 // LeaveAuction : Return the orderbook and market to continuous trading.