+++ risk/mocks/risk_model_mock.go
@@ -5,67 +5,66 @@
 package mocks
 
 import (
+	reflect "reflect"
+
 	types "code.vegaprotocol.io/vega/types"
 	num "code.vegaprotocol.io/vega/types/num"
 	gomock "github.com/golang/mock/gomock"
 	decimal "github.com/shopspring/decimal"
-	reflect "reflect"
-	time "time"
 )
 
-// MockModel is a mock of Model interface
+// MockModel is a mock of Model interface.
 type MockModel struct {
 	ctrl     *gomock.Controller
 	recorder *MockModelMockRecorder
 }
 
-// MockModelMockRecorder is the mock recorder for MockModel
+// MockModelMockRecorder is the mock recorder for MockModel.
 type MockModelMockRecorder struct {
 	mock *MockModel
 }
 
-// NewMockModel creates a new mock instance
+// NewMockModel creates a new mock instance.
 func NewMockModel(ctrl *gomock.Controller) *MockModel {
 	mock := &MockModel{ctrl: ctrl}
 	mock.recorder = &MockModelMockRecorder{mock}
 	return mock
 }
 
-// EXPECT returns an object that allows the caller to indicate expected use
+// EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockModel) EXPECT() *MockModelMockRecorder {
 	return m.recorder
 }
 
-// CalculateRiskFactors mocks base method
-func (m *MockModel) CalculateRiskFactors(arg0 *types.RiskResult) (bool, *types.RiskResult) {
+// CalculateRiskFactors mocks base method.
+func (m *MockModel) CalculateRiskFactors() *types.RiskFactor {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "CalculateRiskFactors", arg0)
-	ret0, _ := ret[0].(bool)
-	ret1, _ := ret[1].(*types.RiskResult)
-	return ret0, ret1
-}
-
-// CalculateRiskFactors indicates an expected call of CalculateRiskFactors
-func (mr *MockModelMockRecorder) CalculateRiskFactors(arg0 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CalculateRiskFactors", reflect.TypeOf((*MockModel)(nil).CalculateRiskFactors), arg0)
-}
-
-// CalculationInterval mocks base method
-func (m *MockModel) CalculationInterval() time.Duration {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "CalculationInterval")
-	ret0, _ := ret[0].(time.Duration)
+	ret := m.ctrl.Call(m, "CalculateRiskFactors")
+	ret0, _ := ret[0].(*types.RiskFactor)
 	return ret0
 }
 
-// CalculationInterval indicates an expected call of CalculationInterval
-func (mr *MockModelMockRecorder) CalculationInterval() *gomock.Call {
+// CalculateRiskFactors indicates an expected call of CalculateRiskFactors.
+func (mr *MockModelMockRecorder) CalculateRiskFactors() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CalculationInterval", reflect.TypeOf((*MockModel)(nil).CalculationInterval))
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CalculateRiskFactors", reflect.TypeOf((*MockModel)(nil).CalculateRiskFactors))
 }
 
-// GetProjectionHorizon mocks base method
+// DefaultRiskFactors mocks base method.
+func (m *MockModel) DefaultRiskFactors() *types.RiskFactor {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "DefaultRiskFactors")
+	ret0, _ := ret[0].(*types.RiskFactor)
+	return ret0
+}
+
+// DefaultRiskFactors indicates an expected call of DefaultRiskFactors.
+func (mr *MockModelMockRecorder) DefaultRiskFactors() *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DefaultRiskFactors", reflect.TypeOf((*MockModel)(nil).DefaultRiskFactors))
+}
+
+// GetProjectionHorizon mocks base method.
 func (m *MockModel) GetProjectionHorizon() decimal.Decimal {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "GetProjectionHorizon")
@@ -73,13 +72,13 @@ func (m *MockModel) GetProjectionHorizon() decimal.Decimal {
 	return ret0
 }
 
-// GetProjectionHorizon indicates an expected call of GetProjectionHorizon
+// GetProjectionHorizon indicates an expected call of GetProjectionHorizon.
 func (mr *MockModelMockRecorder) GetProjectionHorizon() *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetProjectionHorizon", reflect.TypeOf((*MockModel)(nil).GetProjectionHorizon))
 }
 
-// PriceRange mocks base method
+// PriceRange mocks base method.
 func (m *MockModel) PriceRange(arg0, arg1, arg2 decimal.Decimal) (decimal.Decimal, decimal.Decimal) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "PriceRange", arg0, arg1, arg2)
@@ -88,13 +87,13 @@ func (m *MockModel) PriceRange(arg0, arg1, arg2 decimal.Decimal) (decimal.Decima
 	return ret0, ret1
 }
 
-// PriceRange indicates an expected call of PriceRange
+// PriceRange indicates an expected call of PriceRange.
 func (mr *MockModelMockRecorder) PriceRange(arg0, arg1, arg2 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PriceRange", reflect.TypeOf((*MockModel)(nil).PriceRange), arg0, arg1, arg2)
 }
 
-// ProbabilityOfTrading mocks base method
+// ProbabilityOfTrading mocks base method.
 func (m *MockModel) ProbabilityOfTrading(arg0, arg1 *num.Uint, arg2, arg3, arg4 decimal.Decimal, arg5, arg6 bool) decimal.Decimal {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "ProbabilityOfTrading", arg0, arg1, arg2, arg3, arg4, arg5, arg6)
@@ -102,7 +101,7 @@ func (m *MockModel) ProbabilityOfTrading(arg0, arg1 *num.Uint, arg2, arg3, arg4 
 	return ret0
 }
 
-// ProbabilityOfTrading indicates an expected call of ProbabilityOfTrading
+// ProbabilityOfTrading indicates an expected call of ProbabilityOfTrading.
 func (mr *MockModelMockRecorder) ProbabilityOfTrading(arg0, arg1, arg2, arg3, arg4, arg5, arg6 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ProbabilityOfTrading", reflect.TypeOf((*MockModel)(nil).ProbabilityOfTrading), arg0, arg1, arg2, arg3, arg4, arg5, arg6)