+++ statevar/state_variable.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"math/rand"
 	"sort"
+	"sync"
 	"time"
 
 	vegapb "code.vegaprotocol.io/protos/vega"
@@ -23,20 +24,20 @@ const (
 	StateVarConsensusStateCalculationStarted
 	StateVarConsensusStatePerfectMatch
 	StateVarConsensusStateSeekingConsensus
-	StateVarConsensusStateConsensusReached
+	StateVarConsensusStateconsensusReachedLocked
 	StateVarConsensusStateCalculationAborted
 	StateVarConsensusStateError
 	StateVarConsensusStateStale
 )
 
 var stateToName = map[StateVarConsensusState]string{
-	StateVarConsensusStateUnspecified:        "undefined",
-	StateVarConsensusStateCalculationStarted: "consensus_calc_started",
-	StateVarConsensusStatePerfectMatch:       "perfect_match",
-	StateVarConsensusStateSeekingConsensus:   "seeking_consensus",
-	StateVarConsensusStateConsensusReached:   "consensus_reached",
-	StateVarConsensusStateCalculationAborted: "consensus_calc_aborted",
-	StateVarConsensusStateError:              "error",
+	StateVarConsensusStateUnspecified:            "undefined",
+	StateVarConsensusStateCalculationStarted:     "consensus_calc_started",
+	StateVarConsensusStatePerfectMatch:           "perfect_match",
+	StateVarConsensusStateSeekingConsensus:       "seeking_consensus",
+	StateVarConsensusStateconsensusReachedLocked: "consensus_reached",
+	StateVarConsensusStateCalculationAborted:     "consensus_calc_aborted",
+	StateVarConsensusStateError:                  "error",
 }
 
 type StateVariable struct {
@@ -56,6 +57,7 @@ type StateVariable struct {
 	validatorResults            map[string]*statevar.KeyValueBundle // the result of the calculation as received from validators
 	roundsSinceMeaningfulUpdate uint
 	pendingEvents               []pendingEvent
+	lock                        sync.Mutex
 }
 
 func NewStateVar(
@@ -101,21 +103,25 @@ func (sv *StateVariable) GetMarket() string {
 
 // startBlock flushes the pending events.
 func (sv *StateVariable) startBlock(ctx context.Context) {
+	sv.lock.Lock()
 	evts := make([]events.Event, 0, len(sv.pendingEvents))
 	for _, pending := range sv.pendingEvents {
-		evts = append(evts, events.NewStateVarEvent(context.Background(), sv.ID, pending.eventID, pending.state))
+		newEvt := events.NewStateVarEvent(ctx, sv.ID, pending.eventID, pending.state)
+		evts = append(evts, newEvt)
+		protoEvt := newEvt.Proto()
+		sv.log.Info("state-var event sent", logging.String("event", protoEvt.String()))
 	}
+	sv.pendingEvents = []pendingEvent{}
+	sv.lock.Unlock()
 
 	sv.broker.SendBatch(evts)
-	sv.pendingEvents = []pendingEvent{}
 }
 
 // calculation is required for the state variable for the given event id.
 func (sv *StateVariable) eventTriggered(eventID string) error {
-	if sv.log.GetLevel() <= logging.DebugLevel {
-		sv.log.Debug("event triggered", logging.String("state-var", sv.ID), logging.String("event-id", eventID))
-	}
+	sv.lock.Lock()
 
+	sv.log.Info("event triggered", logging.String("state-var", sv.ID), logging.String("event-id", eventID))
 	// if we get a new event while processing an existing event we abort the current calculation and start a new one
 	if sv.eventID != "" {
 		if sv.log.GetLevel() <= logging.DebugLevel {
@@ -128,41 +134,45 @@ func (sv *StateVariable) eventTriggered(eventID string) error {
 			sv.roundsSinceMeaningfulUpdate++
 			if sv.roundsSinceMeaningfulUpdate >= 3 {
 				sv.state = StateVarConsensusStateStale
-				sv.addEvent()
+				sv.addEventLocked()
 			}
 		}
 
 		sv.state = StateVarConsensusStateCalculationAborted
-		sv.addEvent()
+		sv.addEventLocked()
 	}
 
 	// reset any existing state
 	sv.eventID = eventID
 	sv.validatorResults = map[string]*statevar.KeyValueBundle{}
 	sv.state = StateVarConsensusStateCalculationStarted
-	sv.addEvent()
+	sv.addEventLocked()
 
 	if !sv.top.IsValidator() {
+		sv.lock.Unlock()
 		return nil
 	}
 
+	sv.lock.Unlock()
+
 	// kickoff calculation
-	sv.startCalculation(sv.eventID, sv)
+	go sv.startCalculation(sv.eventID, sv)
 
 	return nil
 }
 
 // CalculationFinished is called from the owner when the calculation is completed to kick off consensus.
 func (sv *StateVariable) CalculationFinished(eventID string, result statevar.StateVariableResult, err error) {
+	sv.lock.Lock()
 	if sv.eventID != eventID {
 		sv.log.Warn("ignoring recevied the result of a calculation of an old eventID", logging.String("state-var", sv.ID), logging.String("event-id", eventID))
 	}
-
 	if err != nil {
 		sv.log.Error("could not calculate state for", logging.String("id", sv.ID), logging.String("event-id", eventID))
 		sv.state = StateVarConsensusStateError
-		sv.addEvent()
+		sv.addEventLocked()
 		sv.eventID = ""
+		sv.lock.Unlock()
 		return
 	}
 
@@ -174,6 +184,9 @@ func (sv *StateVariable) CalculationFinished(eventID string, result statevar.Sta
 			Kvb:        sv.converter.InterfaceToBundle(result).ToProto(),
 		},
 	}
+
+	// need to release the lock before we send the transaction command
+	sv.lock.Unlock()
 	sv.cmd.Command(context.Background(), txn.StateVariableProposalCommand, svp, func(err error) { sv.logAndRetry(err, svp) })
 	if sv.log.GetLevel() <= logging.DebugLevel {
 		sv.log.Debug("result calculated and sent to vega", logging.String("validator", sv.top.SelfNodeID()), logging.String("state-var", sv.ID), logging.String("event-id", eventID))
@@ -182,15 +195,25 @@ func (sv *StateVariable) CalculationFinished(eventID string, result statevar.Sta
 
 // logAndRetry logs errors from tendermint transaction submission failure and retries if we're still handling the same event.
 func (sv *StateVariable) logAndRetry(err error, svp *commandspb.StateVariableProposal) {
-	sv.log.Error("failed to send state variable proposal command", logging.String("id", sv.ID), logging.String("event-id", sv.eventID))
+	if err == nil {
+		return
+	}
+	sv.lock.Lock()
+	sv.log.Error("failed to send state variable proposal command", logging.String("id", sv.ID), logging.String("event-id", sv.eventID), logging.Error(err))
 	if svp.Proposal.EventId == sv.eventID {
+		sv.lock.Unlock()
 		sv.log.Info("retrying to send state variable proposal command", logging.String("id", sv.ID), logging.String("event-id", sv.eventID))
 		sv.cmd.Command(context.Background(), txn.StateVariableProposalCommand, svp, func(err error) { sv.logAndRetry(err, svp) })
+		return
 	}
+	sv.lock.Unlock()
 }
 
 // bundleReceived is called when we get a result from another validator corresponding to a given event ID.
 func (sv *StateVariable) bundleReceived(ctx context.Context, t time.Time, node, eventID string, bundle *statevar.KeyValueBundle, rng *rand.Rand, validatorVotesRequired num.Decimal) {
+	sv.lock.Lock()
+	defer sv.lock.Unlock()
+
 	// if the bundle is received for a stale or wrong event, ignore it
 	if sv.eventID != eventID {
 		sv.log.Debug("received a result for a stale event", logging.String("ID", sv.ID), logging.String("from-node", node), logging.String("current-even-id", sv.eventID), logging.String("receivedEventID", eventID))
@@ -207,27 +230,29 @@ func (sv *StateVariable) bundleReceived(ctx context.Context, t time.Time, node, 
 		sv.log.Debug("state var bundle received", logging.String("from-validator", node), logging.String("state-var", sv.ID), logging.String("event-id", eventID))
 	}
 
-	if sv.state == StateVarConsensusStatePerfectMatch || sv.state == StateVarConsensusStateConsensusReached {
+	if sv.state == StateVarConsensusStatePerfectMatch || sv.state == StateVarConsensusStateconsensusReachedLocked {
 		sv.log.Debug("state var bundle received, consensus already reached", logging.String("from-validator", node), logging.String("state-var", sv.ID), logging.String("event-id", eventID))
 		return
 	}
 
 	// save the result from the validator and check if we have a quorum
 	sv.validatorResults[node] = bundle
-	numResults := int64(len(sv.validatorResults))
+	numResults := num.DecimalFromInt64(int64(len(sv.validatorResults)))
 	validatorsNum := num.DecimalFromInt64(int64(len(sv.top.AllNodeIDs())))
-	requiredNumberOfResults := validatorsNum.Mul(validatorVotesRequired).IntPart()
+	requiredNumberOfResults := validatorsNum.Mul(validatorVotesRequired)
 
-	if numResults < requiredNumberOfResults {
+	sv.log.Info("received results for state variable", logging.String("state-var", sv.ID), logging.String("event-id", eventID), logging.Decimal("received", numResults), logging.String("out-of", validatorsNum.String()))
+
+	if numResults.LessThan(requiredNumberOfResults) {
 		if sv.log.GetLevel() <= logging.DebugLevel {
-			sv.log.Debug("waiting for more results for state variable consensus check", logging.String("state-var", sv.ID), logging.String("event-id", eventID), logging.Int64("received", numResults), logging.String("out-of", validatorsNum.String()))
+			sv.log.Debug("waiting for more results for state variable consensus check", logging.String("state-var", sv.ID), logging.String("event-id", eventID), logging.Decimal("received", numResults), logging.String("out-of", validatorsNum.String()))
 		}
 		return
 	}
 
 	// if we're already in seeking consensus state, no point in checking if all match - suffice checking if there's a majority with matching within tolerance
 	if sv.state == StateVarConsensusStateSeekingConsensus {
-		sv.tryConsensus(ctx, t, rng, requiredNumberOfResults)
+		sv.tryConsensusLocked(ctx, t, rng, requiredNumberOfResults)
 		return
 	}
 
@@ -249,7 +274,7 @@ func (sv *StateVariable) bundleReceived(ctx context.Context, t time.Time, node, 
 
 			// initiate a round of voting
 			sv.state = StateVarConsensusStateSeekingConsensus
-			sv.tryConsensus(ctx, t, rng, validatorsNum.Mul(validatorVotesRequired).IntPart())
+			sv.tryConsensusLocked(ctx, t, rng, validatorsNum.Mul(validatorVotesRequired))
 			return
 		}
 	}
@@ -260,11 +285,12 @@ func (sv *StateVariable) bundleReceived(ctx context.Context, t time.Time, node, 
 	}
 	sv.state = StateVarConsensusStatePerfectMatch
 	// convert the result to decimal and let the owner of the state variable know
-	sv.consensusReached(ctx, t, result)
+	sv.consensusReachedLocked(ctx, t, result)
 }
 
 // if the bundles are not all equal to each other, choose one at random and verify that all others are within tolerance.
-func (sv *StateVariable) tryConsensus(ctx context.Context, t time.Time, rng *rand.Rand, requiredMatches int64) {
+// NB: assumes lock has already been acquired.
+func (sv *StateVariable) tryConsensusLocked(ctx context.Context, t time.Time, rng *rand.Rand, requiredMatches num.Decimal) {
 	// sort the node IDs for determinism
 	nodeIDs := make([]string, 0, len(sv.validatorResults))
 	for nodeID := range sv.validatorResults {
@@ -287,9 +313,9 @@ func (sv *StateVariable) tryConsensus(ctx context.Context, t time.Time, rng *ran
 				countMatch++
 			}
 		}
-		if countMatch >= requiredMatches {
-			sv.state = StateVarConsensusStateConsensusReached
-			sv.consensusReached(ctx, t, candidateResult)
+		if num.DecimalFromInt64(countMatch).GreaterThanOrEqual(requiredMatches) {
+			sv.state = StateVarConsensusStateconsensusReachedLocked
+			sv.consensusReachedLocked(ctx, t, candidateResult)
 			return
 		}
 	}
@@ -300,13 +326,16 @@ func (sv *StateVariable) tryConsensus(ctx context.Context, t time.Time, rng *ran
 }
 
 // consensus was reached either through a vote or through perfect matching of all of 2/3 of the validators.
-func (sv *StateVariable) consensusReached(ctx context.Context, t time.Time, acceptedValue *statevar.KeyValueBundle) {
+// NB: assumes lock has already been acquired.
+func (sv *StateVariable) consensusReachedLocked(ctx context.Context, t time.Time, acceptedValue *statevar.KeyValueBundle) {
 	if sv.log.GetLevel() <= logging.DebugLevel {
 		sv.log.Debug("consensus reached", logging.String("state-var", sv.ID), logging.String("event-id", sv.eventID))
 	}
 
 	sv.result(ctx, sv.converter.BundleToInterface(acceptedValue))
-	sv.addEvent()
+	sv.addEventLocked()
+
+	sv.log.Info("consensus reached for state variable", logging.String("state-var", sv.ID), logging.String("event-id", sv.eventID))
 
 	// reset the state
 	sv.eventID = ""
@@ -314,7 +343,9 @@ func (sv *StateVariable) consensusReached(ctx context.Context, t time.Time, acce
 	sv.roundsSinceMeaningfulUpdate = 0
 }
 
-func (sv *StateVariable) addEvent() {
+// addEventLocked adds an event to the pending events.
+// NB: assumes lock has already been acquired.
+func (sv *StateVariable) addEventLocked() {
 	sv.pendingEvents = append(sv.pendingEvents, pendingEvent{sv.eventID, stateToName[sv.state]})
 }