+++ CHANGELOG.md
@@ -19,8 +19,8 @@
 - [4516](https://github.com/vegaprotocol/vega/pull/4516) - Add a command to cleanup all vega node state
 - [4531](https://github.com/vegaprotocol/vega/pull/4531) - Remove Float from network parameters, use `num.Decimal` instead
 - [4537](https://github.com/vegaprotocol/vega/pull/4537) - Send staking asset total supply through consensus
+- [4530](https://github.com/vegaprotocol/vega/pull/4530) - Integrate risk factors with floating point consensus engine
 - [4485](https://github.com/vegaprotocol/vega/pull/4485) - Change snapshot interval default to 1000 blocks
-
 
 ### 🐛 Fixes
 - [4500](https://github.com/vegaprotocol/vega/pull/4500) - Set minimum power to tendermint consensus power for validator